Add sorting to the list of paths before storing them to the golden copy (#56)
* Add validation directory (moved from rms-webtools repo)

* Add command line tool to show opus products output with the given
absolute path of a filespec.

* Update the test framework to sort the list of paths before storing to
the golden copies. With --update param, now the golden copies are
updated only if:
- The golden copy doesn't exists
- The file paths are different than those in the golden copies.

If a list of paths have the same paths but different order than ones
in the golden copy, they will be considered the same. And we won't
update the golden copy in this case.

* Update opus products golden copies, the dictionary keys will be sorted
by 'group' and then 'prority' just like the original opus products
outputs.

* Update opus products golden copies for VG_28xx to have the key sorted
by 'group' and 'priority'.

* Restore opus_products_test by removing line 97

* Update docstrings in pdsfile_test_helper.py to follow Google Python
Style Guide.

---------

Co-authored-by: Robert French <rfrench@seti.org>